Adding new file types for export.
