+++ app/src/main/java/com/canyapan/dietdiaryapp/fragments/ExportFragment.java
@@ -4,22 +4,16 @@ import android.Manifest;
 import android.app.DatePickerDialog;
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.content.res.Resources;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteException;
 import android.net.Uri;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
-import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.ShareCompat;
 import android.support.v7.widget.GridLayout;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -33,24 +27,13 @@ import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.canyapan.dietdiaryapp.Application;
 import com.canyapan.dietdiaryapp.R;
-import com.canyapan.dietdiaryapp.SharingSupportProvider;
 import com.canyapan.dietdiaryapp.db.DatabaseHelper;
-import com.canyapan.dietdiaryapp.db.EventHelper;
 import com.canyapan.dietdiaryapp.helpers.DateTimeHelper;
-import com.canyapan.dietdiaryapp.helpers.ResourcesHelper;
-import com.canyapan.dietdiaryapp.models.Event;
-import com.opencsv.CSVWriter;
 
 import org.joda.time.LocalDate;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
 import java.text.MessageFormat;
-import java.util.concurrent.atomic.AtomicInteger;
 
 public class ExportFragment extends Fragment implements View.OnClickListener {
     public static final String TAG = "ExportFragment";
@@ -59,17 +42,17 @@ public class ExportFragment extends Fragment implements View.OnClickListener {
     private static final String KEY_SELECTED_FORMAT_INT = "FORMAT";
     private static final int REQUEST_EXTERNAL_STORAGE = 30;
 
-    private OnFragmentInteractionListener mListener;
+    protected OnFragmentInteractionListener mListener;
 
-    private GridLayout mGridLayout;
+    protected GridLayout mGridLayout;
     private TextView tvFromDatePicker, tvToDatePicker;
     private Spinner spFormats;
 
-    private LocalDate mFromDate, mToDate;
+    protected LocalDate mFromDate, mToDate;
 
-    private DatabaseHelper mDatabaseHelper;
-    private ProgressDialog mProgressDialog;
-    private ExportAsyncTask mAsyncTask = null;
+    protected DatabaseHelper mDatabaseHelper;
+    protected ProgressDialog mProgressDialog;
+    protected ExportAsyncTask mAsyncTask = null;
 
     public static ExportFragment newInstance() {
         return new ExportFragment();
@@ -164,7 +147,7 @@ public class ExportFragment extends Fragment implements View.OnClickListener {
                 }
 
                 try {
-                    mAsyncTask = (ExportAsyncTask) new ExportAsyncTask(ExportAsyncTask.TO_EXTERNAL).execute();
+                    mAsyncTask = (ExportAsyncTask) new CsvExporter(this, ExportAsyncTask.TO_EXTERNAL).execute();
                 } catch (ExportException e) {
                     Log.e(TAG, "Save to external storage unsuccessful.", e);
                 }
@@ -172,7 +155,7 @@ public class ExportFragment extends Fragment implements View.OnClickListener {
                 return true;
             case R.id.action_share:
                 try {
-                    mAsyncTask = (ExportAsyncTask) new ExportAsyncTask(ExportAsyncTask.TO_SHARE).execute();
+                    mAsyncTask = (ExportAsyncTask) new CsvExporter(this, ExportAsyncTask.TO_SHARE).execute();
                 } catch (ExportException e) {
                     Log.e(TAG, "Share unsuccessful.", e);
                 }
@@ -280,7 +263,7 @@ public class ExportFragment extends Fragment implements View.OnClickListener {
         if (requestCode == REQUEST_EXTERNAL_STORAGE
                 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
             try {
-                mAsyncTask = (ExportAsyncTask) new ExportAsyncTask(ExportAsyncTask.TO_EXTERNAL).execute();
+                mAsyncTask = (ExportAsyncTask) new CsvExporter(this, ExportAsyncTask.TO_EXTERNAL).execute();
             } catch (ExportException e) {
                 Log.e(TAG, "Save to external storage unsuccessful.", e);
             }
@@ -295,213 +278,4 @@ public class ExportFragment extends Fragment implements View.OnClickListener {
         void onShared(Uri uri, LocalDate startDate, LocalDate endDate);
     }
 
-    private class ExportAsyncTask extends AsyncTask<Void, Integer, Boolean> {
-        private static final int TO_EXTERNAL = 0;
-        private static final int TO_SHARE = 1;
-
-        private final File mFile;
-        private final int mDestination;
-        private final AtomicInteger mProgress = new AtomicInteger(0);
-        private String mErrorString = null;
-
-        ExportAsyncTask(int destination) throws ExportException {
-            mDestination = destination;
-
-            String fileName = MessageFormat.format("{0} {1} {2}.csv",
-                    getString(R.string.app_name),
-                    mFromDate.toString("yyyy-MM-dd"),
-                    mToDate.toString("yyyy-MM-dd"));
-
-            switch (destination) {
-                case TO_EXTERNAL:
-                    if (Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
-                        File dir = new File(
-                                Environment.getExternalStorageDirectory(),
-                                Application.APP_DIR);
-                        //noinspection ResultOfMethodCallIgnored
-                        dir.mkdirs();
-
-                        mFile = new File(dir, fileName);
-                    } else {
-                        Log.e(TAG, "SD Card unavailable.");
-                        throw new ExportException(R.string.export_sd_card_unavailable);
-                    }
-                    break;
-                case TO_SHARE:
-                    /*for (File f :
-                            getContext().getCacheDir().listFiles(new FilenameFilter() {
-                                @Override
-                                public boolean accept(File dir, String filename) {
-                                    return filename.startsWith("temp") && filename.endsWith(".csv");
-                                }
-                            })) {
-                        Log.d(TAG, "Deleting temp file " + f.getName());
-                        f.delete();
-                    }*/
-
-                    mFile = new File(getContext().getCacheDir(), fileName);
-                    break;
-                default:
-                    throw new ExportException(R.string.export_unimplemented_destination);
-            }
-        }
-
-        protected void onPreExecute() {
-            // Heads up: This method is also used onCreateView to recreate progress dialog on configuration change.
-            mProgressDialog = new ProgressDialog(getContext());
-            mProgressDialog.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
-            mProgressDialog.setTitle(R.string.export_progress_title);
-            mProgressDialog.setIndeterminate(false);
-            mProgressDialog.setCancelable(false);
-            mProgressDialog.setMax(100);
-            mProgressDialog.show();
-            mProgressDialog.setProgress(mProgress.get());
-        }
-
-        @Override
-        protected Boolean doInBackground(Void... params) {
-            SQLiteDatabase db = null;
-            Cursor cursor = null;
-            CSVWriter csvWriter = null;
-            try {
-                Resources engResources = ResourcesHelper.getEngResources(getContext());
-
-                final OutputStreamWriter os = new OutputStreamWriter(new FileOutputStream(mFile, false), "UTF-8");
-                os.write('\uFEFF'); // Unicode character, U+FEFF BYTE ORDER MARK (BOM) | https://en.wikipedia.org/wiki/Byte_order_mark
-                publishProgress(1);
-
-                csvWriter = new CSVWriter(os);
-
-                csvWriter.writeNext(engResources.getStringArray(R.array.csv_headers));
-
-                db = mDatabaseHelper.getReadableDatabase();
-                cursor = db.query(DatabaseHelper.DBT_EVENT, EventHelper.getDatabaseColumns(),
-                        DatabaseHelper.DBC_EVENT_DATE + " >= ? AND " + DatabaseHelper.DBC_EVENT_DATE + " <= ?",
-                        new String[]{mFromDate.toString(DatabaseHelper.DB_DATE_FORMATTER), mToDate.toString(DatabaseHelper.DB_DATE_FORMATTER)},
-                        null, null, DatabaseHelper.DBC_EVENT_DATE + "," + DatabaseHelper.DBC_EVENT_TIME + "," + DatabaseHelper.DBC_EVENT_ROW_ID);
-
-                publishProgress(5);
-
-                Event model;
-                if (cursor.moveToFirst()) {
-                    int count = cursor.getCount();
-                    Log.d(TAG, MessageFormat.format("There are {0,number,integer} records exporting.", count));
-
-                    String[] types = engResources.getStringArray(R.array.spinner_event_types);
-                    String[] foodTypes = engResources.getStringArray(R.array.spinner_event_food_types);
-                    String[] drinkTypes = engResources.getStringArray(R.array.spinner_event_drink_types);
-
-                    String subType;
-                    int current = 0;
-                    int percent = 0, percent_;
-                    do {
-                        model = EventHelper.parse(cursor);
-
-                        percent_ = (int) Math.floor(++current * 95 / count);
-                        if (percent < percent_) {
-                            percent = percent_;
-                            publishProgress(percent + 5);
-                        }
-
-                        switch (model.getType()) {
-                            case Event.TYPE_FOOD:
-                                subType = foodTypes[model.getSubType()];
-                                break;
-                            case Event.TYPE_DRINK:
-                                subType = drinkTypes[model.getSubType()];
-                                break;
-                            default:
-                                subType = "";
-                        }
-
-                        csvWriter.writeNext(new String[]{
-                                Long.toString(model.getID()),
-                                model.getDate().toString(DatabaseHelper.DB_DATE_FORMATTER),
-                                model.getTime().toString(DatabaseHelper.DB_TIME_FORMATTER),
-                                types[model.getType()],
-                                subType,
-                                model.getDescription()
-                        });
-
-                    } while (cursor.moveToNext());
-                }
-
-                publishProgress(100);
-                return true;
-            } catch (IOException e) {
-                Log.e(TAG, "Content cannot be prepared probably a IO issue.", e);
-                mErrorString = getString(R.string.export_csv_io_exception);
-            } catch (SQLiteException e) {
-                Log.e(TAG, "Content cannot be prepared probably a DB issue.", e);
-                mErrorString = getString(R.string.export_csv_sql_exception);
-            } catch (Exception e) {
-                Log.e(TAG, "Content cannot be prepared.", e);
-                mErrorString = getString(R.string.export_csv_exception);
-            } finally {
-                if (null != csvWriter) {
-                    try {
-                        csvWriter.close();
-                    } catch (IOException ignore) {
-                    }
-                }
-
-                if (null != cursor) {
-                    cursor.close();
-                }
-
-                if (null != db && db.isOpen()) {
-                    db.close();
-                }
-            }
-
-            return false;
-        }
-
-        @Override
-        protected void onProgressUpdate(Integer... values) {
-            if (values.length > 0 && values[0] != null) {
-                mProgress.set(values[0]);
-                mProgressDialog.setProgress(values[0]);
-            }
-        }
-
-        @Override
-        protected void onPostExecute(Boolean result) {
-            mAsyncTask = null;
-
-            if (mProgressDialog.isShowing()) {
-                mProgressDialog.dismiss();
-            }
-
-            if (null == result || result.equals(false)) {
-                Snackbar.make(mGridLayout, mErrorString, Snackbar.LENGTH_INDEFINITE).show();
-            } else if (result.equals(true)) {
-                Snackbar.make(mGridLayout, getString(R.string.export_external_successful, mFile.getName()), Snackbar.LENGTH_SHORT).show();
-
-                if (mDestination == TO_SHARE) {
-                    ShareCompat.IntentBuilder builder = ShareCompat.IntentBuilder.from(getActivity())
-                            .setType(SharingSupportProvider.MIME_TYPE_CSV)
-                            .setStream(Uri.parse(SharingSupportProvider.CONTENT_URI_PREFIX + mFile.getName()));
-
-                    builder.getIntent().addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-
-                    builder.startChooser();
-
-                    mListener.onShared(Uri.fromFile(mFile), mFromDate, mToDate);
-                } else {
-                    mListener.onExported(Uri.fromFile(mFile), mFromDate, mToDate);
-                }
-            }
-        }
-    }
-
-    class ExportException extends Exception {
-        public ExportException(String message) {
-            super(message);
-        }
-
-        public ExportException(@StringRes int message) {
-            this(getString(message));
-        }
-    }
 }
